Optimice la función factorial para que pueda calcular numeros mas grandes, con un limite actual de 20 como entrada

+++ funciones.h
@@ -1,8 +1,9 @@
 #ifndef FUNCIONES_H
 #define FUNCIONES_H
 
-unsigned long int factorial(unsigned short numero) {
-    unsigned long int resultado = 1;
+unsigned long long factorial(unsigned short numero) {
+    // Número de salida maxima: 18,446,744,073,709,551,615
+    unsigned long long resultado = 1;
     while(numero > 0) {
         resultado *= numero;
         numero--;
